seams: implement first version of the add-on
Seams is a Thunderbird add-on to bring some integration with Patchwork,
the patch tracking system, to Thunderbird.

This commit adds a first working version of the add-on, with retrieval
of patch state and relevant metadata from Patchwork.

Note that Linux patchwork is the only supported instance at the moment.

Please refer to the README.md for more details.