popup: provide visual feedback when copying commands
To apply the patch or its series locally, one can click on the commands
displayed in the popup. So far there was no visual feedback to indicate
that copy succeeded (or failed, or that anything at all happened,
really). Provide some feedback by altering the text and color of the
line.