popup: update width for commands when check details are expanded
The width for the commands for applying the patch locally is limited in
order to avoid the popup to be too wide with very little text on its
right side. But when the check details are expanded and widen the popup,
there is no longer a reason for limiting this width. Update it when the
check details are open to avoid a very clumsy effect.