scripts: prepare for supporting multiple instances
Turn the const variables used for hardcoding the pattern string and API
server URL for the kernel patchwrok instance into an array, with a
single element for now.

Instead of searching for a single pattern, search for the pattern of
each Patchwork instance in the array, and return the (first) matching
patchwork instance.