scripts: add support for cover letters
It may be convenient to get the links to a series or to the mailing list
archives for a series from its cover letter, which happens to be at the
root of message threads in Thunderbird.

Add support for cover letters by turning the restriction on " 0/" into a
way to detect cover letters, and by adjusting the Patchwork request
appropriately.

Cover letters have no state, so we clean up the badge. They also have no
checks and it is not relevant to try to apply them by their id, so we
hide all irrelevant sections in the popup when opening it for a cover
letter.