scripts: favour patch/project for which ML corresponds to our instance
When a patch belongs to several projects on a same Patchwork instance,
for example because it was sent to two different mailing lists, each
having its associated project on the instance, then the API request
returns a list of several patches for the message ID, each associated to
one project. Furthermore, the ordering of those projects may differ
between series, which means that for different patches sent to the same
mailing lists (plural), we may display information for different
projects, without having a way to prioritise one.

When selecting the patch from the data returned from the API request,
instead of always picking the first result, let's see if one result
corresponds best to the current Patchwork instance we are working with,
by checking the email address for the mailing list and comparing it to
the pattern we have for the Patchwork instance. This is not a perfect
solution (several projects may match on @vger.kernel.org for example, or
the user may have chosen something different from the email of the
mailing list) but should provide a decent way to pick up the correct
instance in most cases.