linter: add config file and Makefile target for ESLint
Add config file and Makefile target for running the linter ESLint
(https://eslint.org) on the JavaScript files of the add-on.

Use "JavaScript Standard Style" (https://github.com/standard/standard)
with a few modifications:

- Allow dangling commas for some constructs (It avoids touching
  unnecessary lines in Git history when adding new items to arrays or
  objects).
- Reject blocks in curly braces on a single line (Personal preference).
- Use four spaces for indent (Come on! Two spaces, seriously?).
- Use semicolons at the end of statements (To avoid destroying all the
  Git history).
- Use double quotes for strings (Idem).

To run the linter, install ESLint and the "standard" package:

    $ npm install eslint --save-dev
    $ npm install standard --save-dev

DO NOT RUN: Config file initialisation (but we have one already):

    $ npx eslint --init

Then run the linter on the files (or directories)

    $ npx eslint scripts options popup

Alternatives such as yarn are available. For more details, see:
https://eslint.org/docs/user-guide/getting-started