scripts: always clear badge status on new message
So far, the add-on button would not be displayed at all when the action
was disabled because the message was not recognised as a patch. In that
case, we didn't bother cleaning the state of the badge (indicating the
patch status), because it wouldn't show. In Thunderbird 102, the button
is greyed instead of being hidden, and the badge, if existing, is always
shown. Jumping from a patch to a non-patch message results in a grey
button, with a status badge without any sense for the non-patch email.

Fix this by always clearing the badge status when displaying a new
message, instead of doing it only after we have recognised a patch.